+++ server/internal/store/fs/fs_test.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"path/filepath"
 
-	"sourcegraph.com/sourcegraph/rwvfs"
 	"src.sourcegraph.com/sourcegraph/conf"
 
 	"golang.org/x/net/context"
@@ -21,21 +20,12 @@ func testContext() (ctx context.Context, done func()) {
 		panic(err)
 	}
 
-	appdata := filepath.Join(tmpDir, "appdata")
-	err = os.MkdirAll(appdata, 0777)
-	if err != nil {
-		panic(err)
-	}
-
 	ctx = context.Background()
 	ctx = WithReposVFS(ctx, filepath.Join(tmpDir, "repos"))
-	ctx = WithBuildStoreVFS(ctx, rwvfs.Walkable(rwvfs.OS(filepath.Join(tmpDir, "builds"))))
-	ctx = WithDBVFS(ctx, rwvfs.Map(map[string]string{}))
-	ctx = WithAppStorageVFS(ctx, rwvfs.Walkable(rwvfs.OS(appdata)))
 	ctx = conf.WithURL(ctx, &url.URL{Scheme: "http", Host: "example.com"}, nil)
 	return ctx, func() {
 		if err := os.RemoveAll(tmpDir); err != nil {
-			log.Fatalf("Warning: failed to remove fs store temp dir %q: %s.", tmpDir, err)
+			log.Fatalf("Warning: failed to remove temp dir %q: %s.", tmpDir, err)
 		}
 	}
 }